refactoring, again
