+++ frontend/.eslintrc.json
@@ -36,8 +36,11 @@
   },
   "settings": {
     "import/resolver": {
-      "node": {
-        "extensions": [".js", ".jsx", ".ts", ".tsx"]
+      "typescript": {
+        "project": "."
+      },
+      "alias": {
+        "map": [["@/*", "./*"]]
       }
     }
   }
