+++ frontend/components/landing-page/LandingPage.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
-import { Row, Col, Group } from '../styles/Layout'
-import { Heading3, Title } from '../styles/Text'
-import { DesktopWireframes, MobileWireframes } from './Wireframes'
+import { Row, Col, Group } from '@/components/styles/Layout'
+import { Heading3, Title } from '@/components/styles/Text'
+import {
+  DesktopWireframes,
+  MobileWireframes,
+} from '@/components/landing-page/Wireframes'
 
 const LandingPage = () => (
   <>
