+++ frontend/package.json
@@ -39,6 +39,7 @@
     "eslint-config-next": "11.1.2",
     "eslint-config-prettier": "^8.3.0",
     "eslint-config-react-app": "^6.0.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
     "eslint-plugin-flowtype": "^6.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "prettier": "^2.4.1"
