+++ frontend/pages/index.tsx
@@ -1,10 +1,10 @@
 import type { GetServerSidePropsContext } from 'next'
 
-import Header from '../components/header/Header'
-import LandingPage from '../components/landing-page/LandingPage'
-import Dashboard from '../components/dashboard/Dashboard'
-import { AuthUserContext, withAuth } from '../context/auth'
-import { User } from '../types'
+import Header from '@/components/header/Header'
+import LandingPage from '@/components/landing-page/LandingPage'
+import Dashboard from '@/components/dashboard/Dashboard'
+import { AuthUserContext, withAuth } from '@/utils/auth'
+import { User } from '@/utils/types'
 
 interface IndexPageProps {
   user: User
