+++ frontend/pages/polls/[...pid].tsx
@@ -2,18 +2,18 @@ import React, { useState, useCallback } from 'react'
 import { GetServerSidePropsContext } from 'next'
 import { useRouter } from 'next/router'
 
-import { AuthUserContext, withAuth } from '../../context/auth'
-import { doApiRequest } from '../../utils/fetch'
-import { PageType, PollType, Status, User } from '../../types'
-import Nav from '../../components/header/Nav'
-import { Col, Group, Row } from '../../components/styles/Layout'
-import { Button, ToggleButton } from '../../components/styles/Buttons'
-import { Subtitle } from '../../components/styles/Text'
-import { colors } from '../../utils/colors'
-import StatusBar from '../../components/form/StatusBar'
-import PollForm from '../../components/form/poll/PollForm'
-import Preview from '../../components/form/Preview'
-import { convertCamelCase, convertSnakeCase } from '../../utils/utils'
+import { AuthUserContext, withAuth } from '@/utils/auth'
+import { doApiRequest } from '@/utils/fetch'
+import { PageType, PollType, Status, User } from '@/utils/types'
+import Nav from '@/components/header/Nav'
+import { Col, Group, Row } from '@/components/styles/Layout'
+import { Button, ToggleButton } from '@/components/styles/Buttons'
+import { Subtitle } from '@/components/styles/Text'
+import { colors } from '@/components/styles/colors'
+import StatusBar from '@/components/form/StatusBar'
+import PollForm from '@/components/form/poll/PollForm'
+import Preview from '@/components/form/Preview'
+import { convertCamelCase, convertSnakeCase } from '@/utils/utils'
 
 interface iPollPageProps {
   user: User
