--- frontend/utils/gippage.ts
@@ -1,5 +1,0 @@
-import { NextPageContext, NextPage } from 'next'
-
-export type GIPPage<P> = NextPage<P> & {
-  getInitialProps: (ctx: NextPageContext) => P | Promise<P>
-}
