--- frontend/utils/routes.ts
@@ -1,7 +1,0 @@
-export const DEV_STUDENT_LIFE_URL = 'http://localhost:8000'
-export const STUDENT_LIFE_URL = 'https://studentlife.pennlabs.org'
-
-export const DEV_API_URL = 'http://localhost:5000/portal'
-export const API_URL = 'https://api.pennlabs.org'
-
-export const HOME_ROUTE = '/'
